fix(debug-bar): address review findings for the standalone build
- Attach a no-op catch to the debug-bar build promise so a rejection can't
  go unhandled when the main client build throws before the swap-time await
- Select the bundle output by kind === 'entry-point' instead of trusting
  outputs[0]
- Drop the now-unreachable debug-bar exclusion from the BundlesPanel chunk
  walk (the bar is no longer part of the metafile'd build)
- Move formatBuildMessages into its own module, breaking the
  buildDebugBarBundle <-> ComponentRegistry import cycle
